mergedata
merge data changes

+++ src/app/databases/databases.component.ts
@@ -2,6 +2,9 @@ import { Component, OnInit } from '@angular/core';
 import { ApiService } from '../api.service';
 import { DomSanitizer, SafeResourceUrl } from '@angular/platform-browser';
 import { Router } from '@angular/router';
+import { MatDialog , MatDialogConfig} from '@angular/material/dialog';
+// import { MatSnackBar } from '@angular/material/snack-bar';
+import { DeleteCollectionDialogComponent } from '../delete-collection-dialog/delete-collection-dialog.component';
 
 
 
@@ -14,39 +17,49 @@ declare var window: any;
   styleUrls: ['./databases.component.scss'],
 })
 export class DatabasesComponent  {
-  
-
- 
-
- 
   errorMessage: string = '';
   collectionName = '';
   csvMetadata: any = {};
+  deleteMessage: string | null = null;
   
 
   constructor(
     private apiService: ApiService,
     private sanitizer: DomSanitizer,
-    private router: Router
+    private router: Router,private dialog: MatDialog
   ) {}
     
+
   deleteCollection(collectionName: string) {
     if (collectionName) {
-      this.apiService.deleteCollection(collectionName).subscribe(
-        () => {
-          console.log(`Collection ${collectionName} deleted successfully.`);
-          
-        },
-        (error) => {
-          console.error(`Error deleting collection: ${error}`);
-         
+      const dialogRef = this.dialog.open(DeleteCollectionDialogComponent, {
+        data: { collectionName },
+        panelClass: 'custom-dialog-panel' // Add this line
+      });
+  
+      dialogRef.afterClosed().subscribe((result) => {
+        if (result === true) {
+          // User clicked "Yes," proceed with deletion
+          this.apiService.deleteCollection(collectionName).subscribe(
+            () => {
+              const successMessage = `Collection ${collectionName} deleted successfully.`;
+              // Show a success message or take further actions
+            },
+            (error) => {
+              const errorMessage = `Error deleting collection: ${error}`;
+              // Show an error message or take further actions
+            }
+          );
+        } else {
+          // User clicked "No" or closed the dialog, do nothing
         }
-      );
+      });
     } else {
       console.error('Please provide a collection name.');
-      
     }
   }
+  
+
 
 viewTable(collectionName: string): void {
   if (collectionName.trim() === '') {
